+++ charts/core/src/EventMarkers/BaseEventMarker/BaseEventMarker.types.ts
@@ -1,0 +1,17 @@
+import type { Theme } from '@leafygreen-ui/lib';
+
+export const EventLevel = {
+  Warning: 'warning',
+  Info: 'info',
+} as const;
+export type EventLevel = (typeof EventLevel)[keyof typeof EventLevel];
+
+export interface BaseEventMarkerProps {
+  name: string;
+  theme: Theme;
+  label: string;
+  message: string;
+  level: EventLevel;
+  position: [string | number, string | number] | string | number;
+  type: 'line' | 'point';
+}
