+++ charts/core/src/EventMarkers/BaseEventMarker/index.ts
@@ -1,0 +1,2 @@
+export { BaseEventMarker } from './BaseEventMarker';
+export { BaseEventMarkerProps, EventLevel } from './BaseEventMarker.types';
