--- charts/core/src/EventMarkers/EventMarker.types.ts
@@ -1,5 +1,0 @@
-export const EventLevel = {
-  Warning: 'warning',
-  Info: 'info',
-} as const;
-export type EventLevel = (typeof EventLevel)[keyof typeof EventLevel];
